ENG-1020, ENG-958:  Check for dependencies better.
Checks for AWS CLI executable like it does for Docker. (add ref. to Amazon's
instructions)

If Docker is found, but doesn't run properly, check to see if we're running
on a Linux machine; if we are, display note about needing to run as root or
add user to docker group (with ref. to Docker's instructions).